+++ Apps/pisowise-fe/app/projects/[id]/page.tsx
@@ -3,12 +3,12 @@
 import { useParams } from "next/navigation";
 import { useProjectStore } from "@/app/store/projectsPage/projectStore";
 import { useModalStore } from "@/app/store/projectsPage/modalStore";
-import { EditProjectModal } from "../components/modals/EditprojectModal";
-import { ProjectHeader } from "./components/Header";
-import BudgetCard from "./components/cards/BudgetCard";
-import ReceiptsCard from "./components/cards/ReceiptsCard";
-import InsightsCard from "./components/cards/AIInsightsCard";
-import WarningCard from "./components/cards/WarningCard";
+import { EditProjectModal } from "../../../components/projects/modal/EditprojectModal";
+import { ProjectHeader } from "@/components/projects/details/Header";
+import BudgetCard from "../../../components/projects/details/cards/BudgetCard";
+import ReceiptsCard from "@/components/projects/details/cards/ReceiptsCard";
+import InsightsCard from "../../../components/projects/details/cards/AIInsightsCard";
+import WarningCard from "../../../components/projects/details/cards/WarningCard";
 
 export default function ProjectDetailsPage() {
   const params = useParams();
