+++ Apps/pisowise-fe/app/projects/page.tsx
@@ -1,4 +1,4 @@
-import ProjectsPage from "./components/ProjectsPage";
+import ProjectsPage from "@/components/projects/ProjectsPage";
 
 export default function Projects() {
   return (
