+++ Apps/pisowise-fe/components/projects/ProjectsPage.tsx
@@ -4,8 +4,8 @@ import { Plus, SquarePen } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import ProjectCard from "./cards/ProjectCard";
 import CreateProject from "./cards/CreateProjectCard";
-import { CreateProjectModal } from "./modals/CreateProjectModal";
-import { EditProjectModal } from "./modals/EditprojectModal";
+import { CreateProjectModal } from "./modal/CreateProjectModal";
+import { EditProjectModal } from "./modal/EditprojectModal";
 import { useProjectStore } from "@/app/store/projectsPage/projectStore";
 import { useModalStore } from "@/app/store/projectsPage/modalStore";
 import NoProjectsCard from "./cards/NoProjectsCard";
