+++ Apps/pisowise-fe/components/projects/details/cards/AIInsightsCard.tsx
