+++ Apps/pisowise-fe/components/receipts/cards/PurchaseDetails.tsx
