+++ Apps/pisowise-fe/components/receipts/modals/ImageModal.tsx
